- added a real interface for the RepositoryManager
